Update python binding color name calls.
I don't know if this actually works, but the design should be correct.  The basic idea is that we use the 'colorData' call instead of the old C API calls, since the colorData should map directly to a Python dictionary.  (If it doesn't, we need to update colorData (perhaps in SWIG?) so it does.)