Require the 'reactionGlyphIndex' argument.
There were already functions with that argument, and implementation of these functions was broken.  It was also causing me confusion because there was a difference between the 'get' functions and the 'set' functions.  Removing the functions without reactionGlyphIndex seems to have just fixed everything, and made things simpler to boot.